Add unhandled error logging
This adds Reliability Kit's error logging to the uncaught exception
handler provided by Sentry. This is not the most ideal place for this
logging, but trust us we tried to do it differently.

Our hope is that we can do this work properly when we have time to
reconsider our use of Sentry and replace it with a module in Reliability
Kit. For more information on why we can't do this yet, see
[FTDCS-247](https://financialtimes.atlassian.net/browse/FTDCS-247).

Co-Authored-By: Alex Muller <alex.muller@ft.com>
Co-Authored-By: Andy Gout <andy.gout@ft.com>